last minute fixes!
Clean up a few files:

tables.py: added dtype and file path selectors to load_team_mapping and others.

TeamMapping.py - Fixes error with season_year being a string! Forces int.

TeamRevenueView.py - Fixes the error that led to season_year becoming a string.

market_size.csv - save this because I may not be able to find it again.

fan_attendance_revenue.csv - needs to be built before opt 2 is taken. It is created by a later option. For now, I check it in but the tables should preload before the program starts.